code refactoring, minor doc refinements. Added axoness2mesh SSO method. Fixed problem with view normalization druing rendering, glia predictions seem not to be affected. Added time stamp to filehandler logs.
