+++ scripts/WIP/working_w_ssds.py
